rexstan enhancements

+++ pages/user.php
@@ -1,6 +1,6 @@
 <?php
 $func = rex_request('func', 'string');
-$entry_id = rex_request('entry_id', 'int');
+$entry_id = (int) rex_request('entry_id', 'int');
 
 if ('' != filter_input(INPUT_POST, 'newsletter_exportusers')) {
     $func = 'export';
